Rework staging and the watching provider to handle real load
First, don't lock thread while waiting for staging inFile to complete.
Doing Thread.sleep locks the thread, it's better to allow it to
reschedule and do other work. This explains why the single-threaded
watching provider fell apart with real load, and also the
1-Hz-per-thread performance of the polling provider.  As a part of this
the default polling period was modified to sync up with the 1000 ms
GRACE_PERIOD.

Second, split staging in two phases, first to detect that a staged file
exists and second to wait for staging completion.

This allows us to do allocation as soon as the file is being staged
instead of waiting until staging is complete before allocating space for
it.

This also enabled sorting out the Watching provider to actually work
with real load. The main problem here was the use of MODIFY events which
actually reports an event for each write operation to a file in the
watched directory, this essentially produces an event storm at the IO
rates we need in production. To add insult to injury, the Java event
implementation doesn't support the inotify IN_CLOSE_WRITE event so we
need to resort to polling instead to monitor file completion.

While at it, use a timeout when watching events to double-check that
we're in sync with the state of the watched directory since this
mechanism isn't foolproof in some corner cases.

The watching provider has now gained the same threads and period
tunables as the polling provider, but with a smaller period default
since the watching provider uses polling only on files that are in
flight.

There is also a new watchtimeout that defaults to 300 seconds that is
used to trigger the double-check mentioned above.

+++ src/main/java/org/ndgf/endit/StageTask.java
@@ -43,7 +43,7 @@ import static java.util.Arrays.asList;
 import com.google.gson.JsonObject;
 import org.apache.commons.io.FileUtils;
 
-class StageTask implements PollingTask<Set<Checksum>>
+class StageTask implements PollingStageTask<Boolean>
 {
     public static final int ERROR_GRACE_PERIOD = 1000;
 
@@ -52,7 +52,7 @@ class StageTask implements PollingTask<Set<Checksum>>
     private static final int PID = CLibrary.INSTANCE.getpid();
 
     private final static Logger LOGGER = LoggerFactory.getLogger(StageTask.class);
-    
+
     private final Path file;
     private final Path inFile;
     private final Path errorFile;
@@ -60,6 +60,10 @@ class StageTask implements PollingTask<Set<Checksum>>
     private final long size;
     private final String storageClass;
     private final String path;
+    private long delayUntil;
+    private boolean doStart;
+    private boolean doComplete;
+    private boolean doWatch;
 
     StageTask(StageRequest request, Path requestDir, Path inDir)
     {
@@ -72,6 +76,16 @@ class StageTask implements PollingTask<Set<Checksum>>
         requestFile = requestDir.resolve(id);
         storageClass = fileAttributes.getStorageClass();
         path = request.getFileAttributes().getStorageInfo().getMap().get("path");
+        delayUntil = -1;
+        doStart = false;
+        doComplete = false;
+        doWatch = false;
+    }
+
+    @Override
+    public boolean canWatch()
+    {
+        return this.doWatch;
     }
 
     @Override
@@ -80,19 +94,20 @@ class StageTask implements PollingTask<Set<Checksum>>
         return asList(errorFile, inFile);
     }
 
+
+    /* start() returns when inFile exists. */
     @Override
-    public Set<Checksum> start() throws Exception
+    public Boolean start() throws Exception
     {
-        if (Files.isRegularFile(inFile) && Files.size(inFile) == size) {
-            Files.move(inFile, file, StandardCopyOption.ATOMIC_MOVE);
-            return Collections.emptySet();
+        assert !doComplete : "Internal ENDIT provider bug: complete() called before start()";
+
+        doStart = true;
+        doWatch = true;
+
+        if (Files.isRegularFile(inFile)) {
+            return true;
         }
-        else if (Files.isRegularFile(file) && Files.size(file) == size) {
-            // The file already exists at its final destination, no need to
-            // do more...
-            return Collections.emptySet();
-        }
-  
+
         JsonObject jsObj = new JsonObject();
         jsObj.addProperty("file_size", size);
         jsObj.addProperty("parent_pid", PID);
@@ -106,8 +121,24 @@ class StageTask implements PollingTask<Set<Checksum>>
         return null;
     }
 
+    /* complete() completes processing with waiting for a
+     * complete inFile, grace delay for file attributes to be set, and moves
+     * it to outFile.
+     */
     @Override
-    public Set<Checksum> poll() throws IOException, InterruptedException, EnditException
+    public Boolean complete() throws Exception
+    {
+        assert !doStart : "Internal ENDIT provider bug: start() called before complete()";
+
+        doComplete = true;
+
+        return poll();
+    }
+
+
+    /* poll() handles both start() and complete() initiated tasks */
+    @Override
+    public Boolean poll() throws IOException, InterruptedException, EnditException
     {
         if (Files.exists(errorFile)) {
             List<String> lines;
@@ -121,32 +152,40 @@ class StageTask implements PollingTask<Set<Checksum>>
             }
             throw EnditException.create(lines);
         }
-        if (Files.isRegularFile(inFile) && Files.size(inFile) == size) {            
-            Files.deleteIfExists(requestFile);            
-            Thread.sleep(GRACE_PERIOD); 
 
-            // Check if inFile still exists, we've slept for a while and another
-            // thread might have grabbed it. This isn't atomic so we might
-            // still get spurious benign Files.move() errors printed.
-            if(Files.exists(inFile) && Files.size(inFile) == size) {
-                try {
-                    Files.move(inFile, file, StandardCopyOption.ATOMIC_MOVE);
-                } catch (IOException e) {
-                    System.err.println(e);
-                }
-                return Collections.emptySet();
-            }
-            else if(Files.exists(file) && Files.size(file) == size) {
-                // Another thread has already moved the staged file.
-                return Collections.emptySet();
-            }
-            else {
-                // Neither staged inFile or target file exists, error out.
-                List<String> err = List.of("Failed to move staged file in place, neither staged inFile nor target file exists.", "This is likely an ENDIT provider bug and should never happen.");
-                throw EnditException.create(err);
+        if(doStart) {
+            if (Files.isRegularFile(inFile)) {
+                return true;
             }
         }
+        else if(doComplete) {
+            if(Files.isRegularFile(inFile) && Files.size(inFile) == size) {
+                if(delayUntil < 0) {
+                    Files.deleteIfExists(requestFile);
+                    delayUntil = System.currentTimeMillis() + GRACE_PERIOD;
+                    return null;
+                }
+                if(delayUntil > 0 && System.currentTimeMillis() < delayUntil) {
+                    return null;
+                }
+
+                Files.move(inFile, file, StandardCopyOption.ATOMIC_MOVE);
+                return true;
+            }
+        }
+        else {
+            // Neither start() or complete() called.
+            List<String> err = List.of("Internal ENDIT provider bug.", "StageTask: neither start() nor complete() called before poll().");
+            throw EnditException.create(err);
+        }
         return null;
+    }
+
+    @Override
+    public Set<Checksum> checksum() throws Exception
+    {
+        // No proper checksum retention yet.
+        return Collections.emptySet();
     }
 
     @Override
